decompose frontend components
