+++ pyproject.toml
@@ -88,7 +88,7 @@ select = [
     "TID",
     "W",
 ]
-per-file-ignores = { "src/frontend/constants.py" = ["E501"], "src/frontend/frontend.py" = ["E501"], "src/middleware/meta_tag_injection.py" = ["E501"] }
+per-file-ignores = { "src/frontend/components/arena.py" = ["E501"], "src/frontend/components/leaderboard.py" = ["E501"], "src/middleware/meta_tag_injection.py" = ["E501"] }
 
 [tool.ruff.lint.pycodestyle]
 max-line-length = 120
