+++ src/frontend/__init__.py
@@ -1,10 +1,3 @@
-"""
-Frontend package for the Expressive TTS Arena application.
-
-This package contains UI components for the web application, with the main
-Frontend class serving as the entry point for the Gradio interface.
-"""
-
 from .frontend import Frontend
 
 __all__ = ["Frontend"]
