+++ src/frontend/components/__init__.py
@@ -1,0 +1,4 @@
+from .arena import Arena
+from .leaderboard import Leaderboard
+
+__all__ = ["Arena", "Leaderboard"]
